optimized time formats

+++ src/main/java/me/codedred/playtimes/time/formats/OnlyHours.java
@@ -6,7 +6,7 @@ public class OnlyHours implements Timings {
 
     @Override
     public String buildFormat(long time) {
-        return ((double)Math.round((time/3600.0)*100d))/100d + HOURS;
+        return (Math.round((time/3600.0)*100d))/100d + HOURS;
     }
 
 }
